# Speed up `fourier_series`