This PR speeds up the fourier series computation by:
- speeding up the conversion to days since epoch
- pre-allocating the numpy array and optimizing the for loop

```py
%timeit Prophet.fourier_series(DATA["ds"], 7, 3)
```
`main`
169 µs ± 1.38 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

This PR
30.9 µs ± 699 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

Also added in this PR:
- raise error if `series_order` is < 1
- added typing to the method